ENGR00171038 ASRC:bug fix for convertion to 16k
16k output samplerate is not support in our driver and the call of
config a pair IOCTL function would return -1 to program. The test
program should stop the further ASRC operation when -1 is
received.however, because the missing of brackets, the err is given a
wrong value 1(it should be -1).so it will hold the further operation of
ASRC and access some data not defined or so. thus system hang.

Add the missing brackets.

Signed-off-by: Chen Liangjun <b36089@freescale.com>