ENGR00287256 linux-test: Fix the SYNC mode hang issue when connected with MITB.
This patch fixes the hang and crash issue of MLB SYNC mode in the test.
The MITB will casue Sabreauto to hang and crash when testing the SYNC mode.
It is because MITB will cause something error on MLB bus when stopping the
SYNC test. The Sabreauto will keep entering error ISR and hang. Since we
don't know the details about MITB, we call IO_CTRL interface to disable
the interrupt in SYNC mode.

Signed-off-by: Luwei Zhou <b45643@freescale.com